Candidate 4: inline IOUtils into AggregateFiles, drop the junk drawer
IOUtils held three unrelated functions (createDirectoryStructure,
renameFile, printFileContent) only because they all touched the file
system. printFileContent was dead code; the other two were called from a
single site — AggregateFiles.action(). The interface (three functions
with no shared theme) was nearly as complex as the implementation, the
canonical shape of a shallow module.

- Inline createDirectoryStructure as private fun createStagingDirectory in
  AggregateFiles. The Windows POSIX guard from #4 moves with it.
- Inline renameFile as private fun renameFile in AggregateFiles, using
  Kotlin require/check idioms instead of throwing IllegalArgumentException
  / IllegalStateException by hand.
- Delete printFileContent (no callers in src/main or src/test).
- Delete IOUtils.kt and IOUtilsTest.kt. The four old tests were thin
  wrappers around File.mkdirs / File.renameTo — they didn't earn their
  keep against private logic now scoped to one task. Future coverage
  belongs in an AggregateFilesTest that exercises the staging behavior
  end-to-end (see Candidate 2).

The file-system knowledge now lives next to the only place that uses it.
